requirements.txt
PyMySQL==0.9.3
Flask==2.0.3
redis==3.4.1
qrcode==7.3.1
Pillow==9.2.0
numpy==1.23.0
opencv-python==4.6.0.66
